refactor: splitup code into files
